+++ src/components/common/Like.tsx
@@ -1,32 +1,70 @@
 import { icLikeClicked, icLikeHover, icLikeDefault } from 'public/assets/icons';
 import ImageDiv from './ImageDiv';
 import styled, { css } from 'styled-components';
+import { useEffect, useState } from 'react';
+import { api } from '@src/services/api';
 
 interface LikeProps {
   isFromList: boolean;
+  newsId: number;
   isLiked?: boolean;
-  toggleLike: () => void;
 }
 
 function Like(props: LikeProps) {
-  const { isFromList, isLiked, toggleLike } = props;
-  // const [isLikedTemp, setIsLikedTemp] = useState<boolean>(isLiked);
+  const { isFromList, newsId, isLiked } = props;
+  const [token, setToken] = useState('');
+  const [userId, setUserId] = useState('');
+  const [test, setTest] = useState(isLiked);
+  const [updateState, setUpdateState] = useState<number>(0);
 
-  // useEffect(() => {
-  //   setIsLikedTemp(isLiked);
-  // }, []);
+  useEffect(() => {
+    (async () => {
+      const getAccessToken = () => localStorage.getItem('token') ?? '';
+      const getUserId = () => localStorage.getItem('userId') ?? '';
+      if (getAccessToken()) {
+        setToken(getAccessToken());
+      }
+      if (getUserId()) {
+        setUserId(getUserId());
+      }
+    })();
+    setUpdateState(1);
+  }, []);
 
-  console.log('isLiked2', isLiked);
+  useEffect(() => {
+    setTest(isLiked);
+    setUpdateState(0);
+  }, [isLiked, test, updateState]);
+
+  const handleClick = async (newsId: number, test?: boolean) => {
+    if (test) {
+      console.log('데이터 삭제', test);
+      await api.likeService.deleteLikeData({
+        news_id: newsId,
+        access_token: token,
+        user_id: userId,
+      });
+      setTest(false);
+    } else {
+      console.log('데이터 추가', test);
+      await api.likeService.postLikeData({
+        news_id: newsId,
+        access_token: token,
+        user_id: userId,
+      });
+      setTest(true);
+    }
+  };
 
   return (
     <StLike
       type="button"
       onClick={(e) => {
         e.stopPropagation();
-        toggleLike();
+        handleClick(newsId, test);
       }}>
       <StLikeImage isFromList={isFromList}>
-        {isLiked ? (
+        {test ? (
           <ImageDiv className="like" src={icLikeClicked} alt="like" layout="fill" />
         ) : (
           <>
