+++ src/components/common/NewsList.tsx
@@ -15,69 +15,18 @@ interface NewsListProps {
 function NewsList(props: NewsListProps) {
   const { newsList } = props;
   const router = useRouter();
-  // const [likeList, setLikeList] = useState<number[]>([]);
-  const [token, setToken] = useState('');
-  const [userId, setUserId] = useState('');
-
-  const [, setUpdateState] = useState<number>(0);
+  const [, setUpdateState] = useState(0);
 
   useEffect(() => {
     (async () => {
       const { favoriteList } = await api.likeService.getLikeData();
-      // setLikeList(favoriteList.map((like) => like.id));
       console.log('favoriteList', favoriteList);
       newsList.map((news) => {
         news.isLiked = favoriteList.map((like) => like.id).includes(news.id) ? true : false;
       });
-      // console.log('newsList2', newsList);
       setUpdateState(1);
     })();
   }, []);
-
-  // useEffect(() => {
-  //   console.log('likeList', likeList);
-  //   // const tempList = JSON.parse(JSON.stringify(newsList));
-  //   // temp?.map((news: any) => {
-  //   //   news.isLiked = likeList?.includes(news.id) ? true : false;
-  //   // });
-  //   newsList?.map((news) => {
-  //     news.isLiked = likeList?.includes(news.id) ? true : false;
-  //   });
-  //   console.log('newsLisdsft', newsList[0].isLiked);
-  // }, [likeList, newsList]);
-
-  useEffect(() => {
-    (async () => {
-      const getAccessToken = () => localStorage.getItem('token') ?? '';
-      const getUserId = () => localStorage.getItem('userId') ?? '';
-      if (getAccessToken()) {
-        setToken(getAccessToken());
-      }
-      if (getUserId()) {
-        setUserId(getUserId());
-      }
-    })();
-  }, []);
-
-  const handleClick = async (id: number, isLiked?: boolean) => {
-    if (!isLiked) {
-      console.log('데이터 추가 요청');
-      await api.likeService.postLikeData({
-        news_id: id,
-        access_token: token,
-        user_id: userId,
-      });
-    } else {
-      console.log('데이터 삭제 요청');
-      await api.likeService.deleteLikeData({
-        news_id: id,
-        access_token: token,
-        user_id: userId,
-      });
-    }
-  };
-
-  // 버튼 클릭 -> 빈 버튼이었으면 즐찾 추가 (post) -> response를 가지고 화면에 꽉찬 하트 보여줌
 
   return (
     <StNewsList>
@@ -93,8 +42,7 @@ function NewsList(props: NewsListProps) {
                 layout="fill"
                 alt=""
               />
-              {console.log('isLiked', isLiked)}
-              <Like isFromList={true} isLiked={isLiked} toggleLike={() => handleClick(id, isLiked)} />
+              <Like isFromList={true} newsId={id} isLiked={isLiked} />
             </>
           </StThumbnail>
           <StTitle>{title}</StTitle>
